+++ include/anedya_operations.h
@@ -23,6 +23,8 @@
 #define ANEDYA_OP_OTA_UPDATE_STATUS 5
 #define ANEDYA_OP_VALUESTORE_SET 6
 #define ANEDYA_OP_SUBMIT_EVENT 7
+#define ANEDYA_OP_CMD_UPDATE_STATUS 8
+#define ANEDYA_OP_SUBMIT_LOG 9
 
 // Anedya Events
 #define ANEDYA_EVENT_VS_UPDATE_FLOAT 1
@@ -222,9 +224,8 @@ anedya_err_t anedya_op_valuestore_set_bool(anedya_client_t *client, anedya_txn_t
  * This function sends an event to Anedya.
  *
  * @param[in] client Pointer to the `anedya_client_t` structure representing the client.
- * @param[out] txn Pointer to an `anedya_txn_t` structure for the valuestore transaction.
- * @param[in] key A string representing the unique identifier (key) for the value in the valuestore.
- * @param[in] value The floating-point value to be stored in the valuestore.
+ * @param[inout] txn Pointer to an `anedya_txn_t` structure for the valuestore transaction.
+ * @param[in] req_config Pointer to the request config
  *
  * @retval - `ANEDYA_OK` if the valuestore entry is successfully set.
  * @retval - `ANEDYA_ERR_NOT_CONNECTED` if the client is not connected to the server.
@@ -235,6 +236,29 @@ anedya_err_t anedya_op_valuestore_set_bool(anedya_client_t *client, anedya_txn_t
  *          Ensure the appropriate allocation macros are defined.
  */
 anedya_err_t anedya_op_submit_event(anedya_client_t *client, anedya_txn_t *txn, anedya_req_submit_event_t *req_config);
+
+
+/**
+ * @brief Update command status
+ *
+ * This function updates status of a received command to Anedya.
+ *
+ * @param[in] client Pointer to the `anedya_client_t` structure representing the client.
+ * @param[inout] txn Pointer to an `anedya_txn_t` structure for the valuestore transaction.
+ * @param[in] req_config Pointer to the request config
+ *
+ * @retval - `ANEDYA_OK` if the valuestore entry is successfully set.
+ * @retval - `ANEDYA_ERR_NOT_CONNECTED` if the client is not connected to the server.
+ * @retval - Error code if transaction registration or message publishing fails.
+ *
+ * @note Ensure the client is connected before calling this function.
+ * @warning This function uses static or dynamic allocation based on configuration macros.
+ *          Ensure the appropriate allocation macros are defined.
+ */
+anedya_err_t anedya_op_cmd_status_update(anedya_client_t *client, anedya_txn_t *txn, anedya_req_cmd_status_update_t *req_config);
+
+
+anedya_err_t anedya_op_submit_log(anedya_client_t *client, anedya_txn_t *txn, char *log, unsigned int log_len, unsigned long timestamp_ms);
 
 // Reponse handlers
 void _anedya_device_handle_generic_resp(anedya_client_t *client, anedya_txn_t *txn);