+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.8
+
+- Added support for command status update
+- Added log submission support
+- Fix: Datatypes for received command
+
 ## 0.0.7
 
 - Added geo-coordinate data submission support